+++ lib/modules/iconList/IconList.tsx
@@ -1,16 +1,7 @@
 import React, { useMemo } from 'react';
-import {
-  Card,
-  CardContent,
-  CardHeader,
-  List,
-  ListItem,
-  ListItemIcon,
-  ListItemText,
-  Paper,
-  Typography,
-} from '@mui/material';
-import { Source, IconListProps } from './types';
+import { Card, CardContent, CardHeader, List, Paper, Typography } from '@mui/material';
+import { IconListProps } from './types';
+import { IconListListItem } from './IconListListItem.tsx';
 
 const CARD_STYLES = {
   height: '100%',
@@ -32,14 +23,6 @@ const CONTENT_STYLES = {
 
 const LIST_STYLES = {
   padding: 0,
-};
-
-const LIST_ITEM_STYLES = {
-  paddingY: 1.5,
-};
-
-const ICON_STYLES_BASE = {
-  minWidth: 48,
 };
 
 const EMPTY_MESSAGE_STYLES = {
@@ -132,57 +115,5 @@ export const IconList: React.FC<IconListProps> = React.memo(
 );
 
 IconList.displayName = 'IconList';
-
-type Props = {
-  source: Source;
-  showDivider: boolean;
-};
-
-const IconListListItem: React.FC<Props> = React.memo(({ source: source, showDivider }) => {
-  const iconStyles = useMemo(() => {
-    if (source.color) {
-      return {
-        ...ICON_STYLES_BASE,
-        color: source.active ? source.color : 'action.disabled',
-      };
-    }
-
-    return {
-      ...ICON_STYLES_BASE,
-      color: source.active ? 'primary.main' : 'action.disabled',
-    };
-  }, [source.active, source.color]);
-
-  const primaryTextProps = useMemo(
-    () => ({
-      variant: 'body2' as const,
-      fontWeight: source.active ? 'medium' : 'normal',
-    }),
-    [source.active]
-  );
-
-  const secondaryTextProps = useMemo(
-    () => ({
-      variant: 'caption' as const,
-      color: source.active ? 'text.secondary' : 'text.disabled',
-    }),
-    [source.active]
-  );
-
-  const slotProps = useMemo(
-    () => ({
-      primary: primaryTextProps,
-      secondary: secondaryTextProps,
-    }),
-    [primaryTextProps, secondaryTextProps]
-  );
-
-  return (
-    <ListItem divider={showDivider} sx={LIST_ITEM_STYLES}>
-      <ListItemIcon sx={iconStyles}>{source.icon}</ListItemIcon>
-      <ListItemText primary={source.name} secondary={source.description} slotProps={slotProps} />
-    </ListItem>
-  );
-});
 
 IconListListItem.displayName = 'IconListListItem';
